CryptonightR compatibility
Now compatible with Monero v0.14.0.0 which requires the block height
as argument in cn_slow_hash. This is mandatory for the hardfork to block
major_version 10 which brings a new PoW algorithm (CryptonightR).

Block height is now an argument in all calls to the slow hash.

Hashing tests updated to use a block blob with major_version 10.

README now mentioning v0.14.0.0 as minimum Monero version.

circle-ci updated for v0.14.0.0